--- gradle-plugin-integration-test/src/main/proto/com/toasttab/protokt/test/test.proto
@@ -1,10 +1,0 @@
-syntax = "proto3";
-package com.toasttab.protokt.test;
-
-import "protokt/protokt.proto";
-import "google/protobuf/timestamp.proto";
-
-message TestMessage {
-    google.protobuf.Timestamp timestamp = 1 [(.protokt.property).wrap = "java.time.Instant"];
-}
-
